backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
neo4j==5.16.0
GitPython==3.1.40
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.4
tree-sitter-java==0.20.4
# For AI layer (basic RAG setup)
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.5
# Other utilities
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
